+++ src/main/java/com/backend/allreva/module/member/application/MemberService.java
@@ -1,18 +1,22 @@
 package com.backend.allreva.module.member.application;
 
+import com.backend.allreva.common.model.Email;
 import com.backend.allreva.module.concert.artist.application.ArtistService;
 import com.backend.allreva.module.concert.artist.application.dto.ArtistCreateRequest;
 import com.backend.allreva.module.member.application.dto.MemberArtistRequest;
 import com.backend.allreva.module.member.application.dto.MemberDetailResponse;
 import com.backend.allreva.module.member.application.dto.MemberRegisterRequest;
 import com.backend.allreva.module.member.application.dto.NicknameDuplication;
+import com.backend.allreva.module.member.application.dto.OAuthRegisterRequest;
 import com.backend.allreva.module.member.application.dto.RefundAccountRequest;
 import com.backend.allreva.module.member.application.port.MemberDetailRepository;
 import com.backend.allreva.module.member.domain.Member;
 import com.backend.allreva.module.member.domain.MemberRepository;
 import com.backend.allreva.module.member.domain.artist.MemberArtist;
 import com.backend.allreva.module.member.domain.artist.MemberArtistRepository;
+import com.backend.allreva.module.member.domain.value.MemberRole;
 import java.util.List;
+import java.util.UUID;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -35,6 +39,24 @@ public class MemberService {
     public NicknameDuplication isDuplicatedNickname(final String nickname) {
         boolean exists = memberRepository.existsByMemberInfoNickname(nickname);
         return new NicknameDuplication(exists);
+    }
+
+    @Transactional
+    public Member registerByOAuth(final OAuthRegisterRequest request) {
+        Member member = Member.builder()
+                .email(Email.builder().email(request.email()).build())
+                .loginProvider(request.loginProvider())
+                .memberRole(MemberRole.USER)
+                .nickname(generateUniqueNickname())
+                .introduce("")
+                .profileImageUrl(request.profileImageUrl())
+                .build();
+        member.setDefaultRefundAccount();
+        return memberRepository.save(member);
+    }
+
+    private String generateUniqueNickname() {
+        return "user-" + UUID.randomUUID().toString().replace("-", "").substring(0, 8);
     }
 
     @Transactional
