+++ src/main/java/com/backend/allreva/module/member/application/dto/OAuthRegisterRequest.java
@@ -1,0 +1,5 @@
+package com.backend.allreva.module.member.application.dto;
+
+import com.backend.allreva.module.member.domain.value.LoginProvider;
+
+public record OAuthRegisterRequest(String email, LoginProvider loginProvider, String profileImageUrl) {}
